// src/lol/types/gql-static/LolChampionAugmentFragment.ts
/* tslint:disable */
/* eslint-disable */
// @generated
// This file was automatically generated and should not be edited.

// ====================================================
// GraphQL fragment: LolChampionAugmentFragment
// ====================================================

export interface LolChampionAugmentFragment {
  __typename: "ChampionsAugmentsV1DataFlatDto";
  riotId: number | null;
  name: string | null;
  description: string | null;
  rarity: number | null;
}